First attempt at pulling out FitsImageDataSet from HSCDataSet - HSCDataset basically works, tests pass and `hyrax train` on HSC1000 works - FitsImageDataSet has not been tested at all. - rebuild_manifest has not been tested at all. - There are definitely warnings (mostly xcxced) that need to be added to FitsImageDataSet
