+++ .pre-commit-config.yaml
@@ -50,7 +50,7 @@ repos:
       - id: xcxc-check
         name: Check for note-to-self comments (xcxc)
         description: Grep all source files for xcxc which signifies a comment that shouldn't be checked in.
-        entry: bash -c "[[ $(grep -riI xcxc --exclude .pre-commit-config.yaml --exclude-dir _readthedocs --exclude-dir htmlcov ./* >&2 ; echo $?) == 1 ]]"
+        entry: bash -c "[[ $(grep -rniI xcxc --exclude .pre-commit-config.yaml --exclude-dir _readthedocs --exclude-dir htmlcov ./* >&2 ; echo $?) == 1 ]]"
         language: system
         pass_filenames: false
         always_run: true
