+++ src/hyrax/data_sets/__init__.py
@@ -1,4 +1,5 @@
 from .data_set_registry import DATA_SET_REGISTRY, HyraxDataset
+from .fits_image_dataset import FitsImageDataSet
 from .hsc_data_set import HSCDataSet
 from .hyrax_cifar_data_set import HyraxCifarDataSet
 from .inference_dataset import InferenceDataSet
@@ -6,6 +7,7 @@ from .inference_dataset import InferenceDataSet
 __all__ = [
     "DATA_SET_REGISTRY",
     "HyraxCifarDataSet",
+    "FitsImageDataSet",
     "HSCDataSet",
     "InferenceDataSet",
     "Dataset",
